+++ src/ert/__main__.py
@@ -31,16 +31,16 @@ from ert.parsing import ConfigValidationError
 from ert.run_models.multiple_data_assimilation import MultipleDataAssimilation
 from ert.services import StorageService, WebvizErt
 from ert.shared.feature_toggling import FeatureToggling
-from ert.shared.ide.keywords.data.validation_status import ValidationStatus
-from ert.shared.ide.keywords.definitions import (
+from ert.shared.plugins.plugin_manager import ErtPluginContext, ErtPluginManager
+from ert.shared.storage.command import add_parser_options as ert_api_add_parser_options
+from ert.validation import (
     IntegerArgument,
     NumberListStringArgument,
     ProperNameArgument,
     ProperNameFormatArgument,
     RangeStringArgument,
+    ValidationStatus,
 )
-from ert.shared.plugins.plugin_manager import ErtPluginContext, ErtPluginManager
-from ert.shared.storage.command import add_parser_options as ert_api_add_parser_options
 
 
 def run_ert_storage(args: Namespace, _: Optional[ErtPluginManager] = None) -> None:
