+++ src/ert/gui/simulation/ensemble_experiment_panel.py
@@ -13,7 +13,7 @@ from ert.gui.ertwidgets.models.init_iter_value import IterValueModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
 from ert.run_models import EnsembleExperiment
-from ert.shared.ide.keywords.definitions import IntegerArgument, RangeStringArgument
+from ert.validation import IntegerArgument, RangeStringArgument
 
 from .simulation_config_panel import SimulationConfigPanel
