+++ src/ert/gui/simulation/iterated_ensemble_smoother_panel.py
@@ -10,10 +10,7 @@ from ert.gui.ertwidgets.models.targetcasemodel import TargetCaseModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
 from ert.run_models import IteratedEnsembleSmoother
-from ert.shared.ide.keywords.definitions import (
-    ProperNameFormatArgument,
-    RangeStringArgument,
-)
+from ert.validation import ProperNameFormatArgument, RangeStringArgument
 
 from .simulation_config_panel import SimulationConfigPanel
