+++ src/ert/gui/simulation/multiple_data_assimilation_panel.py
@@ -17,7 +17,7 @@ from ert.gui.ertwidgets.models.valuemodel import ValueModel
 from ert.gui.ertwidgets.stringbox import StringBox
 from ert.libres_facade import LibresFacade
 from ert.run_models import MultipleDataAssimilation
-from ert.shared.ide.keywords.definitions import (
+from ert.validation import (
     NumberListStringArgument,
     ProperNameFormatArgument,
     RangeStringArgument,
