--- src/ert/shared/ide/keywords/data/__init__.py
@@ -1,3 +1,0 @@
-from .validation_status import ValidationStatus
-
-__all__ = ["ValidationStatus"]
