+++ src/ert/validation/number_list_string_argument.py
@@ -1,7 +1,7 @@
 import re
 
-from ert.shared.ide.keywords.data.validation_status import ValidationStatus
-from ert.shared.ide.keywords.definitions import ArgumentDefinition
+from .argument_definition import ArgumentDefinition
+from .validation_status import ValidationStatus
 
 
 class NumberListStringArgument(ArgumentDefinition):
