+++ src/ert/validation/range_string_argument.py
@@ -1,8 +1,9 @@
 from typing import Optional
 
 from ert._c_wrappers.config.active_range import ActiveRange
-from ert.shared.ide.keywords.data import ValidationStatus
-from ert.shared.ide.keywords.definitions import ArgumentDefinition
+
+from .argument_definition import ArgumentDefinition
+from .validation_status import ValidationStatus
 
 
 class RangeStringArgument(ArgumentDefinition):
